tool_operate: move config2setopts to separate file, split into subs
To decrease size and complexity.

Closes #17352
